nvim: change render-markdown cmd to call and enable by default
:RenderMarkdown buf_toggle didn't work rly well ???? tf is going on bruh
:RenderMarkdown toggle works every time, but it toggles markdown
rendering for the entire Neovim instance, so if multiple .md buffers
exist, the state will be toggled for all of them.

Also I realized that render-markdown is supposed to provide a nice view
of a markdown file that is also editable, so having the rendering on by
default is fine.