tmux: set some more options
when deleting a window, all windows behind it will be renumbered
e.g. if we have windows 1:foo 2:bar 3:baz and delete win2 bar
the resulting windows will be 1:foo 2:baz

also made it so that splitting a new pane or opening a new window opens
in the cwd of the current pane/window.